Combined SQL for certain tables... read desc.
This is the combined SQL for Event, Event_Info, Park_Images, Parks, and Questions (Amenity and Amenity_Images have incomplete info). We had an Events table but we removed it since it had a lot of redundant info. There shouldn't be anything looking at that table, but changes might need to be made if there are. We have an amenity and amenity_images tables aswell, but they have incomplete information and are in the process of being reworked.