Still some features missing but form connects to database.
Most inputs finished in connection to database when submitting form. Still need to fix on input for file and changing database file datatype to varbinary(max). Lastly to figure what to do with tables that add rows. Then work on back-end feature. Again the sql file is a script for creating all the tables that are connected to the application. You will have to adjust the connection string yourself after running the script that creates the tables.