Fix Progress Bar, Now Layout Broken
Progress Bar is fixed, 8 files required that is for the least. The rest of the file input is ignored from progress bar. Layout is broken because of changing css display attribute to "none" and return it to a viewable form. New idea is to turn a none input html tag to display: none and turn it's children to disable: true.
Then will be working on returning data for applicants that used save and exit.