Fixed the LayOut and Progress Bar Working
Fixed and changed some of the layout. Please read "how to test" to see how it works.
HOW TO TEST: Setup database there is no change in database from previous database script. Run the program. The progress bar will only work when you press the '-' button. An alert message will come up just press enter through it AND the progress bar will update. If any IDs or Names come up, its telling what inputs you are missing and it was used to help me debug javascript. We can turn if off later. The submit button starts black and is disabled. It will be enabled when the progress bar is at 100%.  If you want to submit before finishing form, use the "save and exit" button.